Move input component to domain
WHY: Input component doesn't have any dependency with phaser
